Enhancement: Extract IsTypeTest
