+++ tests/Constraint/ConstraintTest.php
@@ -1114,79 +1114,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintIsType()
-    {
-        $constraint = Assert::isType('string');
-
-        $this->assertFalse($constraint->evaluate(0, '', true));
-        $this->assertTrue($constraint->evaluate('', '', true));
-        $this->assertEquals('is of type "string"', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate(new \stdClass);
-        } catch (ExpectationFailedException $e) {
-            $this->assertStringMatchesFormat(
-                <<<EOF
-Failed asserting that stdClass Object &%x () is of type "string".
-
-EOF
-                ,
-                $this->trimnl(TestFailure::exceptionToString($e))
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintIsType2()
-    {
-        $constraint = Assert::isType('string');
-
-        try {
-            $constraint->evaluate(new \stdClass, 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertStringMatchesFormat(
-                <<<EOF
-custom message
-Failed asserting that stdClass Object &%x () is of type "string".
-
-EOF
-                ,
-                $this->trimnl(TestFailure::exceptionToString($e))
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function resources()
-    {
-        $fh = \fopen(__FILE__, 'r');
-        \fclose($fh);
-
-        return [
-            'open resource'     => [\fopen(__FILE__, 'r')],
-            'closed resource'   => [$fh],
-        ];
-    }
-
-    /**
-     * @dataProvider resources
-     */
-    public function testConstraintIsResourceTypeEvaluatesCorrectlyWithResources($resource)
-    {
-        $constraint = Assert::isType('resource');
-
-        $this->assertTrue($constraint->evaluate($resource, '', true));
-
-        @\fclose($resource);
-    }
-
     public function testConstraintIsNotType()
     {
         $constraint = Assert::logicalNot(
