+++ tests/Constraint/IsTypeTest.php
@@ -1,0 +1,105 @@
+<?php
+/*
+ * This file is part of sebastian/phpunit-framework-constraint.
+ *
+ * (c) Sebastian Bergmann <sebastian@phpunit.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace PHPUnit\Framework\Constraint;
+
+use PHPUnit\Framework\Assert;
+use PHPUnit\Framework\Constraint;
+use PHPUnit\Framework\ExpectationFailedException;
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\TestFailure;
+
+class IsTypeTest extends TestCase
+{
+    public function testConstraintIsType()
+    {
+        $constraint = Assert::isType('string');
+
+        $this->assertFalse($constraint->evaluate(0, '', true));
+        $this->assertTrue($constraint->evaluate('', '', true));
+        $this->assertEquals('is of type "string"', $constraint->toString());
+        $this->assertCount(1, $constraint);
+
+        try {
+            $constraint->evaluate(new \stdClass);
+        } catch (ExpectationFailedException $e) {
+            $this->assertStringMatchesFormat(
+                <<<EOF
+Failed asserting that stdClass Object &%x () is of type "string".
+
+EOF
+                ,
+                $this->trimnl(TestFailure::exceptionToString($e))
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+
+    public function testConstraintIsType2()
+    {
+        $constraint = Assert::isType('string');
+
+        try {
+            $constraint->evaluate(new \stdClass, 'custom message');
+        } catch (ExpectationFailedException $e) {
+            $this->assertStringMatchesFormat(
+                <<<EOF
+custom message
+Failed asserting that stdClass Object &%x () is of type "string".
+
+EOF
+                ,
+                $this->trimnl(TestFailure::exceptionToString($e))
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+
+    /**
+     * @dataProvider resources
+     */
+    public function testConstraintIsResourceTypeEvaluatesCorrectlyWithResources($resource)
+    {
+        $constraint = Assert::isType('resource');
+
+        $this->assertTrue($constraint->evaluate($resource, '', true));
+
+        @\fclose($resource);
+    }
+
+    public function resources()
+    {
+        $fh = \fopen(__FILE__, 'r');
+        \fclose($fh);
+
+        return [
+            'open resource'     => [\fopen(__FILE__, 'r')],
+            'closed resource'   => [$fh],
+        ];
+    }
+
+    /**
+     * Removes spaces in front of newlines
+     *
+     * @param string $string
+     *
+     * @return string
+     */
+    private function trimnl($string)
+    {
+        return \preg_replace('/[ ]*\n/', "\n", $string);
+    }
+}
